real_time_dex_arbitrage_bot/requirements.txt
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
web3==6.20.1
loguru==0.7.2
redis==5.0.8
psycopg[binary]==3.2.10
uvicorn==0.30.6
fastapi==0.115.0
aiohttp==3.10.5
aiofiles==24.1.0
aioSMTPlib==2.0.2
PyYAML==6.0.2
